eks/requirements.txt
appdirs==1.4.4
Arpeggio==1.10.1
attrs==20.3.0
awscli-plugin-endpoint==0.4
backports.functools-lru-cache==1.6.4
boto3==1.17.42
botocore==1.20.42
certifi==2020.11.08
chardet==4.0.0
click==7.1.2
colorama==0.4.3
deepmerge==0.2.1
dill==0.3.3
distlib==0.3.1
docutils==0.15.2
filelock==3.0.12
flake8==3.9.0
Flask==1.1.2
grpcio==1.36.1
hc-aws-at-apple==1.1.2
hc-aws-autotag==0.0.1
hc-aws-eks==1.1.0
hc-aws-rds==1.0.3
himl==0.7.1
hvac==0.10.9
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
logzero==1.6.3
lru-cache==0.2.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
numpy==1.20.1
opencv-python==4.5.1.48
packaging==20.9
parver==0.3.1
pathlib2==2.3.5
pep517==0.10.0
Pillow==8.1.0
pip-tools==6.1.0
pluggy==0.13.1
protobuf==3.15.6
pulumi==2.25.2
pulumi-apple==1.2.4
pulumi-aws==3.38.1
pulumi-cloudinit==0.2.1
pulumi-kubernetes==2.8.4
pulumi-random==3.1.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.2
pytest-mypy==0.8.1
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.5
s3transfer==0.3.6
selenium==3.141.0
semver==2.13.0
simple-photo-gallery==1.4.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.2
Werkzeug==1.0.1
zipp==3.4.1
himl
